Fixed warnings Asn1Object is now an interface Created an AbstractAsn1Object which implements Asn1Object
git-svn-id: https://svn.apache.org/repos/asf/directory/trunks/shared@488299 13f79535-47bb-0310-9956-ffa450edef68

+++ asn1/src/test/java/org/apache/directory/shared/asn1/ber/tlv/ValueTest.java
@@ -63,7 +63,7 @@ public class ValueTest extends TestCase
 
         encoded = Value.getBytes( -27066 );
 
-        Assert.assertEquals( ( byte ) 0x96, ( byte ) encoded[0] );
+        Assert.assertEquals( ( byte ) 0x96, encoded[0] );
         Assert.assertEquals( 0x46, encoded[1] );
 
     }
